Remove oracle submodule
This was an attempt to control the oracle reporting during tests/end-to-end. However, eth-brownie (and dependencies) are quite picky so that both eth-brownie and lido-oracle requirements can be satisfied in a single environment.
Alternative (and closer to reality) approach would be to run oracle in a separate environment on the command line, and have a utility method in "brownie layer" to control it